Divided ong code into submethods of processing_tools submodule
